Consolidate cert requester tests
